+++ CHANGELOG.md
@@ -8,6 +8,16 @@ The format is loosely based on Keep a Changelog.
 
 ---
 
+## v0.15.0 (2026-04-24)
+
+### Feat
+
+- add appearance mode toggle
+
+### Fix
+
+- preserve draft appearance colors
+
 ## v0.14.1 (2026-04-24)
 
 ### Fix